divide unit-test and benchmark-test files
